move inv_adj of existing stock to own folder
