+++ server/service/src/invoice/inventory_adjustment/adjust_existing_stock/insert.rs
@@ -107,7 +107,9 @@ mod test {
     use util::inline_edit;
 
     use crate::{
-        invoice::inventory_adjustment::{insert::InsertInventoryAdjustment, AdjustmentType},
+        invoice::inventory_adjustment::{
+            adjust_existing_stock::InsertInventoryAdjustment, AdjustmentType,
+        },
         service_provider::ServiceProvider,
     };
 
@@ -249,8 +251,11 @@ mod test {
 
     #[actix_rt::test]
     async fn insert_inventory_adjustment_success() {
-        let (_, connection, connection_manager, _) =
-            setup_all("insert_inventory_adjustment_success", MockDataInserts::all()).await;
+        let (_, connection, connection_manager, _) = setup_all(
+            "insert_inventory_adjustment_success",
+            MockDataInserts::all(),
+        )
+        .await;
 
         let service_provider = ServiceProvider::new(connection_manager, "app_data");
         let context = service_provider
