+++ server/service/src/invoice/inventory_adjustment/adjust_existing_stock/mod.rs
