+++ server/service/src/invoice/inventory_adjustment/adjust_existing_stock/validate.rs
@@ -1,7 +1,7 @@
 use repository::{RepositoryError, StockLine, StorageConnection};
 
 use crate::common_stock::{check_stock_line_exists, CommonStockLineError};
-use crate::invoice::inventory_adjustment::insert::AdjustmentType;
+use crate::invoice::inventory_adjustment::adjust_existing_stock::AdjustmentType;
 
 use crate::stocktake_line::validate::{check_active_adjustment_reasons, check_reason_is_valid};
